+++ packages/spacecat-shared-http-utils/src/auth/handlers/scoped-api-key.js
@@ -29,15 +29,18 @@ export default class ScopedApiKeyHandler extends AbstractHandler {
     if (!dataAccess) {
       throw new Error('Data access is required');
     }
+    this.log('Checking for API key in the request headers', 'debug');
 
     const apiKeyFromHeader = headers['x-api-key'];
     if (!hasText(apiKeyFromHeader)) {
       return null;
     }
 
+    this.log(`Checking for API key: ${apiKeyFromHeader}`, 'debug');
     // Keys are stored by their hash, so we need to hash the key to look it up
-    const hashedKey = hashWithSHA256(apiKeyFromHeader);
-    const apiKeyEntity = await dataAccess.getApiKeyByHashedKey(hashedKey);
+    const hashedApiKey = hashWithSHA256(apiKeyFromHeader);
+    this.log(`Checking for API key with hash: ${hashedApiKey}`, 'debug');
+    const apiKeyEntity = await dataAccess.getApiKeyByHashedApiKey(hashedApiKey);
 
     if (!apiKeyEntity) {
       this.log(`No API key entity found in the data layer for the provided API key: ${apiKeyFromHeader}`, 'error');